requirements.txt
Django>=5.0
django-extensions
djangorestframework~=3.16.0
pillow
ipdb
httpie
django-stubs
python-dotenv~=1.1.0
